pagination,message,Image attach

+++ database/migrations/2021_03_29_225637_create_prescriptions_table.php
@@ -17,7 +17,7 @@ class CreatePrescriptionsTable extends Migration
             $table->id('id');
             $table->integer('patient_id');
             $table->integer('quantity');
-            $table->integer('consumption_amount');
+            $table->float('consumption_amount');
             $table->string('department');
             $table->date('date');
             $table->timestamps();
